+++ pkg/unmarshaler/unmarshaler.go
@@ -1,6 +1,7 @@
 package unmarshaler
 
 import (
+	"slices"
 	"strings"
 
 	"github.com/prometheus/common/model"
@@ -16,8 +17,40 @@ type fakeTestFile struct {
 }
 
 type RulesFile struct {
-	Groups       []RuleGroup `yaml:"groups"`
-	fakeTestFile             // Just so we can unmarshal also PromQL test files but ignore them because it has no Groups
+	Groups       GroupsWithComment `yaml:"groups"`
+	fakeTestFile                   // Just so we can unmarshal also PromQL test files but ignore them because it has no Groups
+}
+
+type RulesFileWithComment struct {
+	node           yaml.Node
+	groupsComments []string
+	RulesFile
+}
+
+func (r *RulesFileWithComment) UnmarshalYAML(value *yaml.Node) error {
+	for _, field := range value.Content {
+		if field.Kind == yaml.ScalarNode && field.Value == "groups" {
+			r.groupsComments = strings.Split(field.HeadComment, "\n")
+		}
+	}
+	return unmarshalToNodeAndStruct(value, &r.node, &r.RulesFile)
+}
+
+func (r *RulesFileWithComment) DisabledValidators(commentPrefix string) []string {
+	return disabledValidatorsFromComments(slices.Concat(getYamlNodeComments(r.node, commentPrefix), r.groupsComments), commentPrefix)
+}
+
+type GroupsWithComment struct {
+	node   yaml.Node
+	Groups []RuleGroupWithComment
+}
+
+func (g *GroupsWithComment) UnmarshalYAML(value *yaml.Node) error {
+	return unmarshalToNodeAndStruct(value, &g.node, &g.Groups)
+}
+
+func (g *GroupsWithComment) DisabledValidators(commentPrefix string) []string {
+	return disabledValidatorsFromComments(getYamlNodeComments(g.node, commentPrefix), commentPrefix)
 }
 
 type RuleGroup struct {
@@ -27,6 +60,19 @@ type RuleGroup struct {
 	SourceTenants           []string          `yaml:"source_tenants,omitempty"`
 	Rules                   []RuleWithComment `yaml:"rules"`
 	Limit                   int               `yaml:"limit,omitempty"`
+}
+
+type RuleGroupWithComment struct {
+	node yaml.Node
+	RuleGroup
+}
+
+func (r *RuleGroupWithComment) UnmarshalYAML(value *yaml.Node) error {
+	return unmarshalToNodeAndStruct(value, &r.node, &r.RuleGroup)
+}
+
+func (r *RuleGroupWithComment) DisabledValidators(commentPrefix string) []string {
+	return disabledValidatorsFromComments(getYamlNodeComments(r.node, commentPrefix), commentPrefix)
 }
 
 type RuleWithComment struct {
@@ -46,38 +92,11 @@ func (r *RuleWithComment) OriginalRule() rulefmt.Rule {
 }
 
 func (r *RuleWithComment) UnmarshalYAML(value *yaml.Node) error {
-	err := value.Decode(&r.node)
-	if err != nil {
-		return err
-	}
-	err = value.Decode(&r.rule)
-	if err != nil {
-		return err
-	}
-	return nil
+	return unmarshalToNodeAndStruct(value, &r.node, &r.rule)
 }
 
 func (r *RuleWithComment) DisabledValidators(commentPrefix string) []string {
-	commentPrefix += ":"
-	var disabledValidators []string
-	allComments := strings.Split(r.node.HeadComment, "\n")
-	for _, line := range strings.Split(r.rule.Expr.Value, "\n") {
-		before, comment, found := strings.Cut(line, "#")
-		if !found || strings.TrimSpace(before) != "" {
-			continue
-		}
-		allComments = append(allComments, comment)
-	}
-	for _, comment := range allComments {
-		_, csv, found := strings.Cut(comment, commentPrefix)
-		if !found {
-			continue
-		}
-		validators := strings.Split(csv, ",")
-		for _, v := range validators {
-			vv := strings.TrimSpace(v)
-			disabledValidators = append(disabledValidators, vv)
-		}
-	}
-	return disabledValidators
+	ruleComments := getYamlNodeComments(r.node, commentPrefix)
+	exprComments := getExpressionComments(r.rule.Expr.Value, commentPrefix)
+	return disabledValidatorsFromComments(slices.Concat(ruleComments, exprComments), commentPrefix)
 }