+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added: support for disabling validation using comments in yaml rules per rule group and for the whole file.
+- Docs: Updated/improved documentation on how to disable validations and validation rules.
 
 ## [2.12.0] - 2024-03-07
 - Fixed resolving of the path in `paramsFromFile`. Formerly it was resolved from the current working directory, now it must be a relative path, that will be resolved from the config file location.